+++ include/materials/dielectric.h
@@ -9,12 +9,12 @@
 
 class Dielectric : public Material{ // diffuse material
 public:
-    Dielectric(double index_of_refraction) : index_of_refraction_(index_of_refraction) {}
+    Dielectric(float index_of_refraction) : index_of_refraction_(index_of_refraction) {}
 
     virtual bool scatter(const Ray& ray_in, const HitRecord& hit_record, Color& attenuation, Ray& scattered) const override;
 
-    double index_of_refraction_;
+    float index_of_refraction_;
 
 private:
-    static double reflectance(double cos_theta, double ratio_index_of_refraction);
+    static float reflectance(float cos_theta, float ratio_index_of_refraction);
 };
