+++ include/materials/metal.h
@@ -10,10 +10,10 @@
 
 class Metal : public Material{
 public:
-    Metal(const Color& albedo, double fuzz) : albedo_(albedo), fuzz_(fuzz < 1 ? fuzz : 1.0) {}
+    Metal(const Color& albedo, float fuzz) : albedo_(albedo), fuzz_(fuzz < 1 ? fuzz : 1.0) {}
 
     virtual bool scatter(const Ray& ray_in, const HitRecord& hit_record, Color& attenuation, Ray& scattered) const override;
 
     Color albedo_;
-    double fuzz_;
+    float fuzz_;
 };
